test: move a helper function to the `common.ts` file
