+++ test-utils/common.ts
@@ -8,3 +8,7 @@ export async function setUpFixture<T>(func: () => Promise<T>): Promise<T> {
     return func();
   }
 }
+
+export function createRevertMessageDueToMissingRole(address: string, role: string) {
+  return `AccessControl: account ${address.toLowerCase()} is missing role ${role.toLowerCase()}`;
+}
