--- test-utils/misc.ts
@@ -1,5 +1,0 @@
-function createRevertMessageDueToMissingRole(address: string, role: string) {
-  return `AccessControl: account ${address.toLowerCase()} is missing role ${role.toLowerCase()}`;
-}
-
-export { createRevertMessageDueToMissingRole };
