+++ test/base/RescuableUpgradeable.test.ts
@@ -3,8 +3,7 @@ import { expect } from "chai";
 import { Contract } from "ethers";
 import { HardhatEthersSigner } from "@nomicfoundation/hardhat-ethers/signers";
 import { connect, getAddress, proveTx } from "../../test-utils/eth";
-import { setUpFixture } from "../../test-utils/common";
-import { createRevertMessageDueToMissingRole } from "../../test-utils/misc";
+import { createRevertMessageDueToMissingRole, setUpFixture } from "../../test-utils/common";
 
 describe("Contract 'RescuableUpgradeable'", async () => {
   // Events of the lib contracts
